+++ image++/src/irtkConvolutionWithPadding_2D.cc
@@ -37,7 +37,7 @@ template <class VoxelType> VoxelType irtkConvolutionWithPadding_2D<VoxelType>::G
 
 template <class VoxelType> double irtkConvolutionWithPadding_2D<VoxelType>::Run(int x, int y, int z, int t)
 {
-  float *ptr2;
+  double *ptr2;
   double val, sum;
   int x1, x2, y1, y2;
