+++ image++/src/irtkConvolution_2D.cc
@@ -42,7 +42,7 @@ template <class VoxelType> void irtkConvolution_2D<VoxelType>::SetInput2(irtkGen
 
 template <class VoxelType> double irtkConvolution_2D<VoxelType>::Run(int x, int y, int z, int t)
 {
-  float *ptr2;
+  double *ptr2;
   double val, sum;
   int x1, x2, y1, y2;
